+++ launch/sam.launch.py
@@ -39,6 +39,12 @@ def generate_launch_description():
     is_regular_box_launch_arg = DeclareLaunchArgument(
         "sam_is_regular_box", default_value=TextSubstitution(text="1")
     )
+    encoder_model_file_name_launch_arg = DeclareLaunchArgument(
+        "sam_encoder_model_file_name", default_value=TextSubstitution(text="config/edgesam_encoder_nv12.bin")
+    )
+    decoder_model_file_name_launch_arg = DeclareLaunchArgument(
+        "sam_decoder_model_file_name", default_value=TextSubstitution(text="config/edgesam_decoder_int8_nchw_no_dequantied.bin")
+    )
 
     camera_type = os.getenv('CAM_TYPE')
     print("camera_type is ", camera_type)
@@ -178,6 +184,10 @@ def generate_launch_description():
             {"is_regular_box": LaunchConfiguration(
                 "sam_is_regular_box")},
             {"is_shared_mem_sub": 1},
+            {"encoder_model_file_name": LaunchConfiguration(
+                "sam_encoder_model_file_name")},
+            {"decoder_model_file_name": LaunchConfiguration(
+                "sam_decoder_model_file_name")},
             {"msg_pub_topic_name": LaunchConfiguration(
                 "sam_msg_pub_topic_name")}
         ],
@@ -198,6 +208,8 @@ def generate_launch_description():
             image_height_launch_arg,
             msg_pub_topic_name_launch_arg,
             is_regular_box_launch_arg,
+            encoder_model_file_name_launch_arg,
+            decoder_model_file_name_launch_arg,
             # 启动零拷贝环境配置node
             shared_mem_node,
             # 图片发布pkg
@@ -216,6 +228,8 @@ def generate_launch_description():
             image_height_launch_arg,
             msg_pub_topic_name_launch_arg,
             is_regular_box_launch_arg,
+            encoder_model_file_name_launch_arg,
+            decoder_model_file_name_launch_arg,
             # 启动零拷贝环境配置node
             shared_mem_node,
             # 图片发布pkg