+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for package mobie sam
 
+tros_0.1.1 (2025-06-19)
+------------------
+1. edgesam 新增对多个检测框输入decoder的处理能力。
+2. 新增配置encoder、decoder模型的能力。
+
 tros_0.1.0 (2025-06-13)
 ------------------
 1. 新增edgesam作为模型推理。